requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
asyncio==3.4.3
attrs==25.3.0
certifi==2025.1.31
colorlog==6.9.0
coverage==7.8.0
Cython==3.0.12
frozenlist==1.5.0
grpcio @ file:///home/user0o1/c2-infra/notify/grpc
grpcio-tools==1.71.0
idna==3.10
magic-filter==1.0.12
multidict==6.2.0
mypy-protobuf==3.3.0
pillow==11.1.0
propcache==0.3.1
protobuf==5.29.4
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
setuptools==78.1.0
sliver-py==0.0.19
types-protobuf==5.29.1.20250403
typing_extensions==4.13.0
wheel==0.45.1
yarl==1.18.3
